Bug - Fix error on transpiling using Babel
