--- Server/controllers/index.js
@@ -1,6 +1,0 @@
-const express = require('express')
- 	, router = express.Router();
-
- router.use('/recipes', require('./recipes'));
-
- module.exports = router;
