+++ Server/package.json
@@ -5,18 +5,27 @@
   "main": "app.js",
   "dependencies": {
     "body-parser": "^1.18.2",
+    "chai": "^4.1.2",
+    "chai-http": "^3.0.0",
     "express": "^4.15.5",
+    "mocha": "^3.5.3",
     "morgan": "^1.9.0",
     "nodemon": "^1.12.1"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.2",
+    "babel-polyfill": "^6.26.0",
     "babel-preset-env": "^1.6.0",
-    "babel-preset-es2015": "^6.24.1"
+    "babel-preset-es2015": "^6.24.1",
+    "babel-register": "^6.26.0",
+    "eslint": "^4.7.2"
   },
   "scripts": {
-    "start:dev": "nodemon ./bin/www",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "mocha --compilers js:babel-core/register tests/test.js",
+    "build": "babel ./src -d ./dist",
+    "start": "nodemon ./dist/app.js"
   },
   "keywords": [],
   "author": "",
