+++ Server/src/app.js
@@ -2,23 +2,25 @@
 import logger from 'morgan';
 import bodyParser from 'body-parser';*/
 
-const express = require('express');
-const logger = require('morgan');
-const bodyParser = require('body-parser');
+import express from 'express';
+import logger from 'morgan';
+import bodyParser from 'body-parser';
+import router from './controllers';
 
 // Set up the express app
 const app = express();
-
-// Log requests to the console.
-app.use(logger('dev'));
 
 // Parse incoming requests data (https://github.com/expressjs/body-parser)
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 
-// Setup a default catch-all route that sends back a welcome message in JSON format.
-app.get('*', (req, res) => res.status(200).send({
-  message: 'Welcome to the beginning of nothingness.',
-}));
+// Log requests to the console.
+app.use(logger('dev'));
+app.use('/api/v1/', router);
 
-module.exports = app;
+
+
+// Setup a default catch-all route that sends back a welcome message in JSON format.
+app.listen('8000', () => {
+  console.log("The More-Recipe App Server is running");
+});
